Move the math utilities from quaternion -> utilities by creating its header and src files
